classifier/requirements.txt
matplotlib~=3.2.1
torch~=1.7.0
scikit-learn~=0.24.0rc1
tqdm~=4.46.1
numpy~=1.19.0
Pillow~=7.1.2
torchvision~=0.8.2
scipy~=1.4.1
transformers~=2.11.0
pandas~=1.1.5
gensim~=3.8.3
termcolor~=1.1.0